+++ examples/RSC/app/page.tsx
@@ -1,6 +1,7 @@
 import { serverStorage } from "@/storage.server";
 import ServerUserInfo from "./user-info.server"
-import { StateVocabProvider } from "@yakocloud/state-vocab/server";
+
+const { StateVocabProvider } = serverStorage
 
 export default async function Home() {
   return (
@@ -9,17 +10,23 @@ export default async function Home() {
       <p>Values set via <code>getState</code> on the server:</p>
 
       <StateVocabProvider
-        value={serverStorage.set({
+        value={{
           user: {
             name: "Name",
             role: "Role",
+            id: 1,
+            payload: {
+              hash: "abcd",
+              userId: 1,
+              email: "test@test.tst",
+            }
           },
           person: {
             address: {
               city: "NY"
             }
           }
-        })}
+        }}
       >
         <ServerUserInfo />
       </StateVocabProvider>