+++ examples/RSC/app/user-info.server.tsx
@@ -3,11 +3,12 @@ import ClientUserInfo from "./user-info"
 
 export default async function ServerUserInfo() {
   const name = serverStorage.user.name.getState()
-  console.log("!!! name", name)
 
   const role = serverStorage.user.role.getState()
-  console.log("!!! role", role)
   return (
-    <ClientUserInfo />
+    <>
+      {name} - {role}
+      <ClientUserInfo />
+    </>
   )
 }