+++ examples/RSC/storage.ts
@@ -5,6 +5,14 @@ export const storage = setupStorage({
   user: {
     name: defineState<string>(),
     role: defineState<string>(),
+    id: defineState({
+      defaultValue: -1
+    }),
+    payload: defineState<{
+      hash: string
+      userId: number
+      email?: string
+    }>()
   },
   person: {
     address: {