+++ src/client.ts
@@ -1,4 +1,3 @@
 "use client"
 
 export { clientify } from "./setup.client"
-export { StateVocabClientProvider as StateVocabProvider } from "./provider.client";