+++ src/server.ts
@@ -1,2 +1,1 @@
-export { StateVocabProvider } from "./provider"
 export { serverify } from "./setup.server"